+++ modules/actions/change_preset.js
@@ -5,7 +5,18 @@ export function actionChangePreset(entityID, oldPreset, newPreset, skipFieldDefa
         var tags = entity.tags;
 
         // preserve tags that the new preset might care about, if any
-        if (oldPreset) tags = oldPreset.unsetTags(tags, geometry, newPreset && newPreset.addTags ? Object.keys(newPreset.addTags) : null);
+        var preserveKeys;
+        if (newPreset) {
+            preserveKeys = [];
+            if (newPreset.addTags) {
+                preserveKeys = preserveKeys.concat(Object.keys(newPreset.addTags));
+            }
+            newPreset.fields().concat(newPreset.moreFields())
+                .filter(f => f.matchGeometry(geometry))
+                .map(f => f.key).filter(Boolean)
+                .forEach(key => preserveKeys.push(key));
+        }
+        if (oldPreset) tags = oldPreset.unsetTags(tags, geometry, preserveKeys);
         if (newPreset) tags = newPreset.setTags(tags, geometry, skipFieldDefaults);
 
         return graph.replace(entity.update({tags: tags}));