+++ CHANGELOG.md
@@ -46,6 +46,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 #### :bug: Bugfixes
 * Fix selection of best background source when starting on a zoomed-out view ([#9325])
 * Fix leaking of localized strings in combo fields when taginfo service is unavailable ([#9342])
+* Keep tags when changing presets if the new preset has a field for it ([#9341], [#9104])
 #### :rocket: Presets
 * Support tagging schema v5 ([#9320]):
   * Add new field type `colour` ([schema-builder#38], [#8782])
@@ -56,9 +57,11 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Make settings like CDN and API urls configurable in a central config file (`/config/is.js`)
 
 [#8105]: https://github.com/openstreetmap/iD/issues/8105
+[#9104]: https://github.com/openstreetmap/iD/issues/9104
 [#9294]: https://github.com/openstreetmap/iD/issues/9294
 [#9320]: https://github.com/openstreetmap/iD/pull/9320
 [#9325]: https://github.com/openstreetmap/iD/issues/9325
+[#9341]: https://github.com/openstreetmap/iD/issues/9341
 [#9342]: https://github.com/openstreetmap/iD/issues/9342
 [schema-builder#38]: https://github.com/ideditor/schema-builder/pull/38